aufs: tiny, add a lock for debugging
A new debug-able rw_semaphore reports a problem in reinit_br_wh().
But it is NOT an actual problem, since Big Aufs Lock is already held and
locking sb->s_root after that is meaningless.
Here I follow the message from debug macros, but it may be changed in
the future.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>