aufs: dynop, lifetime is maintained by branch object
The dynop objects are referenced and used by inode or file objects.
But they are kept in a module global list. Getting the operation object
from the branch fs and generating the operation of for aufs starts the
life of dynop object. Its referece count is incremented by branch
object. Only the branch object increments and decrements the
counter. When a branch is removed from aufs, it decrements the counter.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>